% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{AR}
\alias{AR}
\title{Constructor for AR(p) time series}
\usage{
AR(ar_params, start_values, n, sd = 1)
}
\arguments{
\item{ar_params}{Vector of AR parameters, which should have length \code{p}. Parameters must be numeric values.}

\item{start_values}{Vector containing the \code{p} start values.}

\item{n}{A length-one numeric vector specifying the length of the time series one wants to obtain. Should be greater or equal to \code{p}.}

\item{sd}{Standard deviation for generating the white noise components (length-one numeric vector).}
}
\value{
The value returned is an \code{AR} object, having the length and start values specified in beforehand.
}
\description{
Function which can be used to generate an AR(p) time series.
}
\details{
The white noise components will be drawn independently from a centered normal distribution with standard deviation \code{sd}.
}
\examples{
AR(ar_params = c(0.3, 0.7), start_values = c(1,2), n = 30, sd = 1)

}
