% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DurbinLevinson.R
\name{DL_predictor}
\alias{DL_predictor}
\title{Predictor based on the \code{Durbin-Levinson} algorithm}
\usage{
DL_predictor(ts_obj, pred_len = 1, entire_ts = TRUE)
}
\arguments{
\item{ts_obj}{A stationary time series, must be a \code{TimeSeries} class.}

\item{pred_len}{Number of steps one wants to predict.}

\item{entire_ts}{Logical vector, determines whether the original time series with predictions appended (TRUE) or only the predictions (FALSE) is returned.}
}
\value{
The return value is a \code{TimeSeries} object. Depending on the choice of \code{entire_ts}, we either obtain both the original time series with appended predictions or only the predictions made by the algorithm.
}
\description{
Takes a \code{TimeSeries} object and predicts a specified amount of steps. Computations are based on the \code{Durbin-Levinson} algorithm.
}
\details{
This algorithm utilizes the sample autocovariance function \code{ACF} as estimator for the autocovariance.
}
\examples{
ar_ts <- AR(ar_params = 0.5, start_values = 1, n = 50, sd = 1)
DL_predictor(ar_ts, pred_len=5, entire_ts = FALSE)

}
