% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InnovationsAlgorithm.R
\name{innovations_predictor}
\alias{innovations_predictor}
\title{Predictor based on the \code{Innovations} algorithm}
\usage{
innovations_predictor(ts_obj, pred_len = 1, entire_ts = TRUE)
}
\arguments{
\item{ts_obj}{A stationary time series, must be a \code{TimeSeries} class.}

\item{pred_len}{Number of steps one wants to predict, must be length-one numeric.}

\item{entire_ts}{Logical value, which determins whether the original time series with predictions appended (TRUE) or only the predictions (FALSE) is returned.}
}
\value{
The return value is a \code{TimeSeries} object. Depending on the choice of \code{entire_ts}, we either obtain both the original time series with appended predictions or only the predictions made by the algorithm.
}
\description{
Takes a \code{TimeSeries} object and predicts a specified amount of steps. Computations are based on the \code{Innovations} algorithm.
}
\details{
This algorithm utilizes the sample autocovariance function \code{ACF} as estimator for the autocovariance.
}
\examples{
ma_ts <- MA(ma_params = 0.5, n = 50, sd = 1)
innovations_predictor(ma_ts, pred_len=5, entire_ts = FALSE)

}
\references{
Peter J. Brockwell, Richard A. Davis (2016) \emph{Introduction to Time-Series Analysis and Forecasting}; Springer Verlag.
}
