% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{MA-class}
\alias{MA-class}
\title{A S4 class to represent a MA time series}
\description{
Child class of \code{TimeSeries}, used for MA(q) time series. \code{q} refers to the length of \code{ma_params}.
}
\section{Slots}{

\describe{
\item{\code{ma_params}}{Numeric vector containing the parameters for generating an MA time series.}
}}

