% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{MA}
\alias{MA}
\title{Constructor for MA(q) time series}
\usage{
MA(ma_params, n, sd = 1)
}
\arguments{
\item{ma_params}{Vector of MA parameters, which should have length \code{q}. Parameters must be numeric values.}

\item{n}{A length-one numeric vector specifying the length of the time series one wants to obtain. Should be greater or equal to \code{p}.}

\item{sd}{Standard deviation for generating the white noise components (length-one numeric vector).}
}
\value{
The value returned is a \code{MA} object, having the length specified in beforehand.
}
\description{
Function which can be used to generate an MA(q) time series.
}
\details{
The white noise components will be drawn independently from a centered normal distribution with standard deviation \code{sd}.
}
\examples{
MA(ma_params = c(0.5, 0.7), n = 40, sd = 2)

}
