% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodogram.R
\name{periodogram}
\alias{periodogram}
\title{Periodogram for a time series}
\usage{
periodogram(ts_obj)
}
\arguments{
\item{ts_obj}{A time series, must be a \code{TimeSeries} class.}
}
\value{
The return value is a list with values for the frequency and the spectral density.
}
\description{
Takes a \code{TimeSeries} object and computes the periodogram of the time series.
}
\examples{
ar_time_series <- AR(start_values = c(1, 1), ar_params = c(0.6, 0.6), sd = 1, n = 100)
periodogram(ar_time_series)
}
\references{
Peter J. Brockwell, Richard A. Davis (2016) \emph{Introduction to Time-Series Analysis and Forecasting}; Springer Verlag.
}
