% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_periodogram}
\alias{plot_periodogram}
\title{Plot function for the periodogram}
\usage{
plot_periodogram(ts_obj)
}
\arguments{
\item{ts_obj}{A time series, must be a \code{TimeSeries} class.}
}
\value{
The return value is a plot of the periodogram of a time series.
}
\description{
The function plots the periodogram of a given time series.
}
\examples{
ar_time_series <- AR(start_values = c(1, 1), ar_params = c(0.6, 0.6), sd = 1, n = 100)
plot_periodogram(ar_time_series)
ar_time_series_2 <- AR(ar_params = c(1,-0.9), n = 100, start_values = c(1,0.1))
plot_periodogram(ar_time_series_2)

}
