% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_SACVF}
\alias{plot_SACVF}
\title{Plotting the SACVF or SACF}
\usage{
plot_SACVF(ts_obj, acf = FALSE, max_lag = NULL)
}
\arguments{
\item{ts_obj}{A time series, must be a \code{TimeSeries} class.}

\item{acf}{logical. Plots the SACVF if FALSE, SACF otherwise .}

\item{max_lag}{maximum lag at which to calculate the SACVF/SACF .}
}
\value{
The return value is a plot of the SACF/SACVF of a time series.
}
\description{
The function plots the SACVF/SACF of a given time series.
}
\examples{
ar_time_series <- AR(start_values = c(1, 1), ar_params = c(0.6, 0.6), sd = 1, n = 100)
plot_SACVF(ar_time_series,acf=FALSE,max_lag=20)

}
