% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_ts_overview}
\alias{plot_ts_overview}
\title{TimeSeries Overview}
\usage{
plot_ts_overview(ts_obj)
}
\arguments{
\item{ts_obj}{A time series, must be a \code{TimeSeries} class.}
}
\value{
The return value is a composition plot consisting of three or four plots (depending on the availability of the spectral density).
}
\description{
The function plots a the timeseries, its SACVF, spectral density(if available) and periodogram.
}
\examples{
ar_time_series <- AR(start_values = c(1, 1), ar_params = c(0.6, 0.6), sd = 1, n = 100)
plot_periodogram(ar_time_series)
ar_time_series_2 <- AR(ar_params = c(1,-0.9), n = 100, start_values = c(1,0.1))
plot_periodogram(ar_time_series_2)

}
