% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{resample}
\alias{resample}
\alias{resample,AR-method}
\alias{resample,AR}
\alias{resample,MA-method}
\alias{resample,MA}
\title{Resample function for a time series}
\usage{
resample(ts_obj)

\S4method{resample}{AR}(ts_obj)

\S4method{resample}{MA}(ts_obj)
}
\arguments{
\item{ts_obj}{A time series, must be a \code{TimeSeries} class.}
}
\value{
The value returned is an \code{TimeSeries} object, having the same length, start values, parameters and standard deviation as the origin time series.
}
\description{
Function which can be used to resample a time series.
}
\examples{
ar_time_series <- AR(ar_params = c(0.3, 0.7), start_values = c(1,2), n = 30, sd = 1)
resample(ar_time_series)
ma_time_series <- MA(ma_params = c(0.5, 0.7), n = 40, sd = 2)
resample(ma_time_series)
}
