% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleAutoCovarianceFunction.R
\name{sampleACVF}
\alias{sampleACVF}
\title{Sample autocovariance function}
\usage{
sampleACVF(ts_obj, h)
}
\arguments{
\item{ts_obj}{A time series, which must be a \code{TimeSeries} class.}

\item{h}{Point for which we want the function value of the sample autocovariance function, which must be a length-one numeric vector.}
}
\value{
The return value is a numeric value which can be regarded as estimation of the autocovariance function.
}
\description{
This function takes a \code{TimeSeries} object and computes the sample autocovariance function (SACVF) of the data.
}
\examples{
ar_ts <- AR(ar_params = 0.5, start_values = 1, n = 50, sd = 1)
sampleACVF(ar_ts,2)
}
\references{
Peter J. Brockwell, Richard A. Davis (2016) \emph{Introduction to Time-Series Analysis and Forecasting}; Springer Verlag.
}
