% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectralDensity.R
\name{spectral_density}
\alias{spectral_density}
\title{Function returning the spectral density function for a given AR/MA time series}
\usage{
spectral_density(ts_obj)
}
\arguments{
\item{ts_obj}{A time series, must be a \code{TimeSeries} class.}
}
\value{
The return value is the spectral density function for the given TimeSeries object.
}
\description{
Takes a \code{TimeSeries} object. If TimeSeries is of \code{AR} or \code{MA} class, its corresponding spectral density function is computed using the AR/MA parameters. Otherwise an error message is produced.
}
\details{
The returned spectral density function is defined for frequencies in the interval from 0 to 0.5. For each given frequency, the spectral density function computes the corresponding spectral density.
}
\examples{
AR1 <- AR(ar_params = 0.5, start_values = 1, n = 50, sd = 1)
spect_dens <- spectral_density(AR1)
spect_dens(0.2)
}
\references{
Peter J. Brockwell, Richard A. Davis (2016) \emph{Introduction to Time-Series Analysis and Forecasting}; Springer Verlag.
}
