% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{TimeSeries-class}
\alias{TimeSeries-class}
\title{A S4 class to represent a time series}
\description{
A S4 class to represent a time series
}
\section{Slots}{

\describe{
\item{\code{sd}}{Standard deviation used for generating the white noise components (length-one numeric vector).}

\item{\code{n}}{Length of the time series (length-one numeric vector).}

\item{\code{data}}{Values of the time series (numeric vector with length equal to \code{n}).}
}}

