% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{vec_to_ts}
\alias{vec_to_ts}
\title{Vector to time series}
\usage{
vec_to_ts(vec)
}
\arguments{
\item{vec}{A numeric vector.}
}
\value{
The value returned is an \code{TimeSeries} object, having the same length as the vector.
}
\description{
Function which transforms a vector into a \code{TimeSeries} object.
}
\examples{
vec <- 1:10
vec_to_ts(vec)

}
