% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeitreihenR-package.R
\docType{package}
\name{ZeitreihenR-package}
\alias{ZeitreihenR}
\alias{ZeitreihenR-package}
\title{ZeitreihenR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Aaron Mäkel \email{cv234@stud.uni-heidelberg.de}

Authors:
\itemize{
  \item Eivind Hjertholm \email{xxx@stud.uni-heidelberg.de}
  \item Victoria Mezger \email{victoria.mezger@stud.uni-heidelberg.de}
  \item Julia Renner \email{xxx@stud.uni-heidelberg.de}
  \item Klaus Stier \email{xxx@stud.uni-heidelberg.de}
}

}
\keyword{internal}
