% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeitreihenR-package.R
\docType{package}
\name{ZeitreihenR-package}
\alias{ZeitreihenR}
\alias{ZeitreihenR-package}
\title{ZeitreihenR: Package for constructing TimeSeries objects, using different algorithms and visualisations.}
\description{
The package contains S4-classes as well as methods for time series and predictions for them. The definitions and methods are based on the book *Introduction to Time-Series Analysis and Forecasting* by Peter J. Brockwell and Richard A. Davis.
}
\author{
\strong{Maintainer}: Aaron Mäkel \email{cv234@stud.uni-heidelberg.de}

Authors:
\itemize{
  \item Eivind Hjertholm \email{xxx@stud.uni-heidelberg.de}
  \item Victoria Mezger \email{victoria.mezger@stud.uni-heidelberg.de}
  \item Julia Renner \email{xxx@stud.uni-heidelberg.de}
  \item Klaus Stier \email{xxx@stud.uni-heidelberg.de}
}

}
\keyword{internal}
